- Add warning dialog component and module - Refactor book page - Refactor home page - Use modal window instead of alert()
A new WarningDialog component was added to the application to provide more user-friendly warning messages. Additionally, the visibility of user roles has been updated both in the user component and in the authentication checks in the app.component. The Book Reviews and Book Ratings sections in the book component have been merged into a single 'Reviews and Ratings' section. Lastly, minor UI adjustments and code cleanup were made throughout the application.